Pass fp32 beta logits to the XPU KDA kernel
`kda_attention` applies the sigmoid to the beta logits itself and reads
them as float32, but `b_proj` runs in the model dtype, so the XPU branch
handed the kernel a bfloat16 tensor and tripped its dtype validation
before any work was done.

+++ vllm/model_executor/layers/mamba/gdn/kimi_gdn_linear_attn.py
@@ -367,7 +367,9 @@ class KimiGatedDeltaNetAttention(GatedDeltaNetAttention):
                 k,
                 v,
                 g1,
-                beta,
+                # The kernel applies the sigmoid itself but reads the logits as
+                # float32, whereas the projection runs in the model dtype.
+                beta.float(),
                 self.prefix,
             )
             core_attn_out = self.o_norm(core_attn_out, g2)
